LayoutTests/fast/masking/parsing-clip-path-shape-expected.txt

Test that clip-path shapes accept different length units

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS innerStyle("-webkit-clip-path", "circle(0 at 0 0)") is "circle(0px at 0% 0%)"
PASS innerStyle("-webkit-clip-path", "circle(1px at -1px +1px)") is "circle(1px at -1px 1px)"
PASS innerStyle("-webkit-clip-path", "circle(1.5px at -1.5px +1.5px)") is "circle(1.5px at -1.5px 1.5px)"
PASS innerStyle("-webkit-clip-path", "circle(.5px at -.5px +.5px)") is "circle(0.5px at -0.5px 0.5px)"
PASS innerStyle("-webkit-clip-path", "circle(1em at -1em +1em)") is "circle(1em at -1em 1em)"
PASS innerStyle("-webkit-clip-path", "circle(1.5em at -1.5em +1.5em)") is "circle(1.5em at -1.5em 1.5em)"
PASS innerStyle("-webkit-clip-path", "circle(.5em at -.5em +.5em)") is "circle(0.5em at -0.5em 0.5em)"
PASS innerStyle("-webkit-clip-path", "circle(1ex at 1ex 1ex)") is "circle(1ex at 1ex 1ex)"
PASS innerStyle("-webkit-clip-path", "circle(1rem at 1rem 1rem)") is "circle(1rem at 1rem 1rem)"
PASS computedStyle("-webkit-clip-path", "circle(1.5em at .5em 1em)") is "circle(12px at 4px 8px)"
PASS computedStyle("-webkit-clip-path", "circle(1.5rem at .5rem 1rem)") is "circle(24px at 8px 16px)"
PASS innerStyle("-webkit-clip-path", "circle(1vw at 1vw 1vw)") is "circle(1vw at 1vw 1vw)"
PASS innerStyle("-webkit-clip-path", "circle(1vh at 1vh 1vh)") is "circle(1vh at 1vh 1vh)"
PASS innerStyle("-webkit-clip-path", "circle(1vmin at 1vmin 1vmin)") is "circle(1vmin at 1vmin 1vmin)"
PASS computedStyle("-webkit-clip-path", "circle(1.5vw at .5vw 1vw)") is "circle(1.5vw at 0.5vw 1vw)"
PASS computedStyle("-webkit-clip-path", "circle(1.5vh at .5vh 1vh)") is "circle(1.5vh at 0.5vh 1vh)"
PASS computedStyle("-webkit-clip-path", "circle(1.5vmin at .5vmin 1vmin)") is "circle(1.5vmin at 0.5vmin 1vmin)"
PASS computedStyle("-webkit-clip-path", "circle(150% at 50% 100%)") is "circle(150% at 50% 100%)"
PASS computedStyle("-webkit-clip-path", "inset(45% 45% 90% 60% round 25% 10%)") is "inset(45% 45% 90% 60% round 25% 10% 25% 10% / 25% 10% 25% 10%)"
PASS computedStyle("-webkit-clip-path", "ellipse(100% 100% at 100% 100%)") is "ellipse(100% 100% at 100% 100%)"
PASS computedStyle("-webkit-clip-path", "polygon(10% 20%, 30% 40%, 40% 50%)") is "polygon(10% 20%, 30% 40%, 40% 50%)"
PASS innerStyle("-webkit-clip-path", "content-box") is "content-box"
PASS innerStyle("-webkit-clip-path", "padding-box") is "padding-box"
PASS innerStyle("-webkit-clip-path", "border-box") is "border-box"
PASS innerStyle("-webkit-clip-path", "margin-box") is "margin-box"
PASS innerStyle("-webkit-clip-path", "fill") is "fill"
PASS innerStyle("-webkit-clip-path", "stroke") is "stroke"
PASS innerStyle("-webkit-clip-path", "view-box") is "view-box"
PASS innerStyle("-webkit-clip-path", "circle(50% at 50% 50%) content-box") is "circle(50% at 50% 50%) content-box"
PASS innerStyle("-webkit-clip-path", "circle(50% at 50% 50%) padding-box") is "circle(50% at 50% 50%) padding-box"
PASS innerStyle("-webkit-clip-path", "margin-box circle(50% at 50% 50%)") is "margin-box circle(50% at 50% 50%)"
PASS innerStyle("-webkit-clip-path", "border-box circle(50% at 50% 50%)") is "border-box circle(50% at 50% 50%)"
PASS innerStyle("-webkit-clip-path", "circle(50% at 50% 50%) view-box") is "circle(50% at 50% 50%) view-box"
PASS computedStyle("-webkit-clip-path", "content-box") is "content-box"
PASS computedStyle("-webkit-clip-path", "padding-box") is "padding-box"
PASS computedStyle("-webkit-clip-path", "border-box") is "border-box"
PASS computedStyle("-webkit-clip-path", "margin-box") is "margin-box"
PASS computedStyle("-webkit-clip-path", "fill") is "fill"
PASS computedStyle("-webkit-clip-path", "stroke") is "stroke"
PASS computedStyle("-webkit-clip-path", "view-box") is "view-box"
PASS computedStyle("-webkit-clip-path", "circle(50% at 50% 50%) content-box") is "circle(50% at 50% 50%) content-box"
PASS computedStyle("-webkit-clip-path", "circle(50% at 50% 50%) padding-box") is "circle(50% at 50% 50%) padding-box"
PASS computedStyle("-webkit-clip-path", "margin-box circle(50% at 50% 50%)") is "circle(50% at 50% 50%) margin-box"
PASS computedStyle("-webkit-clip-path", "border-box circle(50% at 50% 50%)") is "circle(50% at 50% 50%) border-box"
PASS computedStyle("-webkit-clip-path", "circle(50% at 50% 50%) fill") is "circle(50% at 50% 50%) fill"
PASS innerStyle("-webkit-clip-path", "circle(1 at 1px 1px)") is null
PASS computedStyle("-webkit-clip-path", "circle(1 at 1px 1px)") is "none"
PASS innerStyle("-webkit-clip-path", "circle(px at 1px 1px)") is null
PASS computedStyle("-webkit-clip-path", "circle(px at 1px 1px)") is "none"
PASS innerStyle("-webkit-clip-path", "circle(1p at 1px 1px)") is null
PASS computedStyle("-webkit-clip-path", "circle(1p at 1px 1px)") is "none"
PASS innerStyle("-webkit-clip-path", "circle(calc() at 1px 1px)") is null
PASS computedStyle("-webkit-clip-path", "circle(calc() at 1px 1px)") is "none"
PASS innerStyle("-webkit-clip-path", "circle(-1.5px at -1.5px +1.5px)") is null
PASS computedStyle("-webkit-clip-path", "circle(-1.5px at -1.5px +1.5px)") is "none"
PASS innerStyle("-webkit-clip-path", "inset(1cm 1mm 1in 1px round -1pt, 1pc)") is null
PASS computedStyle("-webkit-clip-path", "inset(1cm 1mm 1in 1px round -1pt, 1pc)") is "none"
PASS innerStyle("-webkit-clip-path", "inset(1cm 1mm 1in 1px round 1pt -1pc)") is null
PASS computedStyle("-webkit-clip-path", "inset(1cm 1mm 1in 1px round 1pt -1pc)") is "none"
PASS innerStyle("-webkit-clip-path", "ellipse(-1em 1em at 1em 1em)") is null
PASS computedStyle("-webkit-clip-path", "ellipse(-1em 1em at 1em 1em)") is "none"
PASS innerStyle("-webkit-clip-path", "ellipse(1em -1em at 1em 1em)") is null
PASS computedStyle("-webkit-clip-path", "ellipse(1em -1em at 1em 1em)") is "none"
PASS innerStyle("-webkit-clip-path", "fill circle(50% at 50% 50%) content-box") is null
PASS computedStyle("-webkit-clip-path", "fill circle(50% at 50% 50%) content-box") is "none"
PASS innerStyle("-webkit-clip-path", "margin-box padding-box") is null
PASS computedStyle("-webkit-clip-path", "margin-box padding-box") is "none"
PASS innerStyle("-webkit-clip-path", "padding-box border-box circle(50% at 50% 50%)") is null
PASS computedStyle("-webkit-clip-path", "padding-box border-box circle(50% at 50% 50%)") is "none"
PASS innerStyle("-webkit-clip-path", "circle(50% at 50% 50%) content-box stroke") is null
PASS computedStyle("-webkit-clip-path", "circle(50% at 50% 50%) content-box stroke") is "none"
PASS innerStyle("-webkit-clip-path", "content-box margin-box padding-box") is null
PASS computedStyle("-webkit-clip-path", "content-box margin-box padding-box") is "none"
PASS innerStyle("-webkit-clip-path", "none content-box") is null
PASS computedStyle("-webkit-clip-path", "none content-box") is "none"
PASS innerStyle("-webkit-clip-path", "url(#test) padding-box") is null
PASS computedStyle("-webkit-clip-path", "url(#test) padding-box") is "none"
PASS innerStyle("-webkit-clip-path", "polygon(0, 0)") is null
PASS computedStyle("-webkit-clip-path", "polygon(0, 0)") is "none"
PASS innerStyle("-webkit-clip-path", "polygon(0 0, 0)") is null
PASS computedStyle("-webkit-clip-path", "polygon(0 0, 0)") is "none"
PASS innerStyle("-webkit-clip-path", "polygon(0)") is null
PASS computedStyle("-webkit-clip-path", "polygon(0)") is "none"
PASS innerStyle("-webkit-clip-path", "polygon()") is null
PASS computedStyle("-webkit-clip-path", "polygon()") is "none"
PASS innerStyle("-webkit-clip-path", "polygon(evenodd)") is null
PASS computedStyle("-webkit-clip-path", "polygon(evenodd)") is "none"
PASS innerStyle("-webkit-clip-path", "polygon(nonzero)") is null
PASS computedStyle("-webkit-clip-path", "polygon(nonzero)") is "none"
PASS successfullyParsed is true

TEST COMPLETE
